Add power select and data reset
Several improvements in this version. Added a data reset if the controller loses connection (so as not to lose control of the car), added MAX throttle toggle using the right joystick press to control low (indoor), medium (outdoor), and high (want to break the car?) throttle settings. Various cleanups and memory improvement also.